requirements.txt
openai==2.8.1
python-dotenv==1.2.1
snowflake-snowpark-python==1.42.0
snowflake-ml-python==1.20.0
streamlit==1.52.0